Moved hornet configs
